Move Store to separate module
