+++ cinescout/account/tmdb/data/build.gradle.kts
@@ -9,6 +9,7 @@ moduleDependencies {
             domain()
         }
     }
+    store()
     utils {
         kotlin()
     }
