+++ cinescout/movies/data/build.gradle.kts
@@ -6,6 +6,7 @@ moduleDependencies {
     movies {
         domain()
     }
+    store()
     utils {
         kotlin()
     }
