+++ cinescout/movies/data/remote/tmdb/src/commonTest/kotlin/cinescout/movies/data/remote/tmdb/RealTmdbMovieDataSourceTest.kt
@@ -19,12 +19,12 @@ import cinescout.movies.domain.testdata.MovieTestData
 import cinescout.movies.domain.testdata.MovieWithDetailsTestData
 import cinescout.movies.domain.testdata.MovieWithPersonalRatingTestData
 import cinescout.movies.domain.testdata.TmdbMovieIdTestData
-import cinescout.store.pagedDataOf
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
+import store.builder.pagedDataOf
 import kotlin.test.Test
 import kotlin.test.assertEquals
