+++ cinescout/movies/data/remote/trakt/src/commonTest/kotlin/cinescout/movies/data/remote/trakt/RealTraktMovieDataSourceTest.kt
@@ -7,12 +7,12 @@ import cinescout.movies.data.remote.trakt.service.TraktMovieService
 import cinescout.movies.data.remote.trakt.testdata.GetRatingsTestData
 import cinescout.movies.domain.model.Rating
 import cinescout.movies.domain.testdata.MovieTestData
-import cinescout.store.pagedDataOf
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
+import store.builder.pagedDataOf
 import kotlin.test.Test
 import kotlin.test.assertEquals
