+++ cinescout/movies/data/src/commonTest/kotlin/cinescout/movies/data/RealMovieRepositoryTest.kt
@@ -10,8 +10,6 @@ import cinescout.movies.domain.testdata.MovieKeywordsTestData
 import cinescout.movies.domain.testdata.MovieTestData
 import cinescout.movies.domain.testdata.MovieWithDetailsTestData
 import cinescout.movies.domain.testdata.MovieWithPersonalRatingTestData
-import cinescout.store.Paging
-import cinescout.store.toPagedData
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.coVerifySequence
@@ -20,6 +18,8 @@ import io.mockk.mockk
 import io.mockk.verify
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.test.runTest
+import store.Paging
+import store.builder.toPagedData
 import kotlin.test.Test
 import kotlin.test.assertEquals
