+++ cinescout/movies/domain/src/commonMain/kotlin/cinescout/movies/domain/usecase/GetAllWatchlistMovies.kt
@@ -2,8 +2,8 @@ package cinescout.movies.domain.usecase
 
 import cinescout.movies.domain.MovieRepository
 import cinescout.movies.domain.model.Movie
-import cinescout.store.PagedStore
-import cinescout.store.Paging
+import store.PagedStore
+import store.Paging
 
 class GetAllWatchlistMovies(
     private val movieRepository: MovieRepository
