+++ cinescout/movies/domain/src/commonMain/kotlin/cinescout/movies/domain/usecase/GetMovieCredits.kt
@@ -5,8 +5,8 @@ import cinescout.error.DataError
 import cinescout.movies.domain.MovieRepository
 import cinescout.movies.domain.model.MovieCredits
 import cinescout.movies.domain.model.TmdbMovieId
-import cinescout.store.Refresh
 import kotlinx.coroutines.flow.Flow
+import store.Refresh
 
 class GetMovieCredits(
     private val movieRepository: MovieRepository
