+++ cinescout/network/trakt/build.gradle.kts
@@ -4,6 +4,7 @@ plugins {
 
 moduleDependencies {
     network()
+    store()
     utils {
         kotlin()
     }
