+++ cinescout/store/build.gradle.kts
@@ -1,22 +1,18 @@
 plugins {
     id("cinescout.kotlin")
-    alias(libs.plugins.kotlin.serialization)
-}
-
-moduleDependencies {
-    store()
-    utils {
-        kotlin()
-    }
-}
-
-dependencies {
-    commonMainImplementation(libs.bundles.base)
-    commonMainImplementation(libs.kotlin.serialization.json)
-
-    commonTestImplementation(libs.bundles.test.kotlin)
 }
 
 kotlin {
     jvm()
+}
+
+moduleDependencies {
+    test.kotlin()
+    utils.kotlin()
+}
+
+dependencies {
+    commonMainImplementation(libs.bundles.base)
+
+    commonTestImplementation(libs.bundles.test.kotlin)
 }
