+++ cinescout/store/src/commonMain/kotlin/store/DefaultRefreshInterval.kt
@@ -1,0 +1,6 @@
+package store
+
+import kotlin.time.DurationUnit.MINUTES
+import kotlin.time.toDuration
+
+val DefaultRefreshInterval = 3.toDuration(MINUTES)
