+++ cinescout/store/src/commonMain/kotlin/store/PagedStore.kt
@@ -1,4 +1,4 @@
-package cinescout.store
+package store
 
 import arrow.core.Either
 import arrow.core.continuations.either
@@ -6,41 +6,16 @@ import arrow.core.left
 import arrow.core.right
 import cinescout.error.DataError
 import cinescout.error.NetworkError
-import cinescout.utils.kotlin.DataRefreshInterval
-import cinescout.utils.kotlin.ticker
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.combineTransform
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.distinctUntilChangedBy
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.flow.transform
-
-/**
- * Creates a flow that combines Local data and Remote data.
- * First emit Local data, only if available, then emits Local data updated from Remote or Remote errors
- *
- * @param refresh:
- *  * If [Refresh.Once] refresh each [DataRefreshInterval]
- *  * If [Refresh.IfNeeded] refresh only if there is no local data
- *  * If [Refresh.Never] do not refresh
- *  Default is [Refresh.Once]
- * @param fetch lambda that returns Remote data
- * @param read lambda that returns a Flow of Local data
- * @param write lambda that saves Remote data to Local
- */
-fun <T> Store(
-    refresh: Refresh = Refresh.Once,
-    fetch: suspend () -> Either<NetworkError, T>,
-    read: () -> Flow<Either<DataError.Local, T>>,
-    write: suspend (T) -> Unit
-): Store<T> = StoreImpl(buildStoreFlow(refresh, fetch, read, write))
+import store.builder.toPagedData
 
 /**
  * Creates a flow that combines Local data and Remote data, when remote data is paged
@@ -113,8 +88,6 @@ fun <T, B, PI : Paging.Page, PO : Paging> PagedStore(
     )
 }
 
-interface Store<T> : Flow<Either<DataError, T>>
-
 interface PagedStore<T, P : Paging> : Store<PagedData<T, P>> {
 
     fun filterIntermediatePages(): Flow<Either<DataError, PagedData<T, P>>> =
@@ -133,80 +106,34 @@ interface PagedStore<T, P : Paging> : Store<PagedData<T, P>> {
     fun loadMore(): PagedStore<T, P>
 }
 
-fun <T, R> Store<T>.map(
-    transform: (Either<DataError, T>) -> Either<DataError.Remote, R>
-): Store<R> = with(this as StoreImpl<T>) {
-    StoreImpl(flow.map(transform))
-}
+internal class PagedStoreImpl<T, P : Paging>(
+    internal val flow: Flow<Either<DataError, PagedData<T, P>>>,
+    internal val onLoadMore: () -> Unit,
+    internal val onLoadAll: () -> Unit
+) : PagedStore<T, P>, Flow<Either<DataError, PagedData<T, P>>> by flow {
 
-fun <T, R, P : Paging> PagedStore<T, P>.map(
-    transform: (Either<DataError, PagedData<T, P>>) ->
-    Either<DataError.Remote, PagedData<R, P>>
-): PagedStore<R, P> = with(this as PagedStoreImpl<T, P>) {
-    PagedStoreImpl(flow.map(transform), onLoadMore, onLoadAll)
-}
-
-fun <T, P : Paging> PagedStore<T, P>.distinctUntilDataChanged(): PagedStore<T, P> =
-    with(this as PagedStoreImpl<T, P>) {
-        PagedStoreImpl(
-            flow.distinctUntilChangedBy { either ->
-                either.map { pagedData ->
-                    pagedData.data
+    override suspend fun getAll(): Either<DataError, List<T>> {
+        onLoadAll()
+        return this.transform { either ->
+            either.tap { pagedData ->
+                if (pagedData.isLastPage()) {
+                    emit(pagedData.data.right())
                 }
-            },
-            onLoadMore,
-            onLoadAll
-        )
+            }.tapLeft { error ->
+                emit(error.left())
+            }
+        }.first()
     }
 
-private fun <T> buildStoreFlow(
-    refresh: Refresh,
-    fetch: suspend () -> Either<NetworkError, T>,
-    read: () -> Flow<Either<DataError.Local, T>>,
-    write: suspend (T) -> Unit
-): Flow<Either<DataError, T>> {
-    val remoteFlow = when (refresh) {
-        Refresh.IfNeeded -> flow {
-            emit(null)
-            read().first().tapLeft {
-                val remoteDataEither = fetch()
-                remoteDataEither.tap { remoteData ->
-                    write(remoteData)
-                }
-                emit(ConsumableData.of(remoteDataEither))
-            }
-        }
-        Refresh.Never -> flowOf(null)
-        Refresh.Once -> flow {
-            emit(null)
-            val remoteDataEither = fetch()
-            remoteDataEither.tap { remoteData ->
-                write(remoteData)
-            }
-            emit(ConsumableData.of(remoteDataEither))
-        }
-        is Refresh.WithInterval -> ticker<ConsumableData<T>?>(refresh.interval) {
-            val remoteDataEither = fetch()
-            remoteDataEither.tap { remoteData ->
-                write(remoteData)
-            }
-            emit(ConsumableData.of(remoteDataEither))
-        }.onStart { emit(null) }
+    override fun loadAll(): PagedStore<T, P> {
+        onLoadAll()
+        return this
     }
-    return combineTransform(
-        remoteFlow,
-        read()
-    ) { consumableRemoteData, localEither ->
-        consumableRemoteData?.consume { remoteEither ->
-            val remote = remoteEither.mapLeft { networkError ->
-                DataError.Remote(networkError = networkError)
-            }
-            emit(remote)
-        }
-        localEither
-            .tap { local -> emit(local.right()) }
-            .tapLeft { localError -> if (refresh is Refresh.Never) emit(localError.left()) }
-    }.distinctUntilChanged()
+
+    override fun loadMore(): PagedStore<T, P> {
+        onLoadMore()
+        return this
+    }
 }
 
 private fun <T, B, PI : Paging.Page, PO : Paging> buildPagedStoreFlow(
@@ -246,39 +173,3 @@ private fun <T, B, PI : Paging.Page, PO : Paging> buildPagedStoreFlow(
             localEither.tap { local -> emit(local.right() as Either<DataError.Remote, PagedData<T, PO>>) }
         }
     }
-
-
-internal class StoreImpl<T> ( internal val flow: Flow<Either<DataError, T>>) :
-    Store<T>, Flow<Either<DataError, T>> by flow
-
-
-internal class PagedStoreImpl<T, P : Paging>(
-    internal val flow: Flow<Either<DataError, PagedData<T, P>>>,
-    internal val onLoadMore: () -> Unit,
-    internal val onLoadAll: () -> Unit
-) : PagedStore<T, P>, Flow<Either<DataError, PagedData<T, P>>> by flow {
-
-    override suspend fun getAll(): Either<DataError, List<T>> {
-        onLoadAll()
-        return this.transform { either ->
-            either.tap { pagedData ->
-                if (pagedData.isLastPage()) {
-                    emit(pagedData.data.right())
-                }
-            }.tapLeft { error ->
-                emit(error.left())
-                // emit(DataError.Remote(networkError = error.networkError).left())
-            }
-        }.first()
-    }
-
-    override fun loadAll(): PagedStore<T, P> {
-        onLoadAll()
-        return this
-    }
-
-    override fun loadMore(): PagedStore<T, P> {
-        onLoadMore()
-        return this
-    }
-}
