+++ cinescout/store/src/commonMain/kotlin/store/ext/StoreExtensions.kt
@@ -1,0 +1,13 @@
+package store.ext
+
+import arrow.core.Either
+import cinescout.error.DataError
+import kotlinx.coroutines.flow.map
+import store.Store
+import store.StoreImpl
+
+fun <T, R> Store<T>.map(
+    transform: (Either<DataError, T>) -> Either<DataError.Remote, R>
+): Store<R> = with(this as StoreImpl<T>) {
+    StoreImpl(flow.map(transform))
+}
