+++ cinescout/suggestions/domain/build.gradle.kts
@@ -7,6 +7,7 @@ moduleDependencies {
     movies {
         domain()
     }
+    store()
     utils {
         kotlin()
     }
