--- cinescout/utils/kotlin/src/commonMain/kotlin/cinescout/utils/kotlin/DataRefreshInterval.kt
@@ -1,6 +1,0 @@
-package cinescout.utils.kotlin
-
-import kotlin.time.DurationUnit.MINUTES
-import kotlin.time.toDuration
-
-val DataRefreshInterval = 3.toDuration(MINUTES)
